+++ sdk/core/core-tracing/src/instrumenter.ts
@@ -9,7 +9,7 @@ import {
 } from "./interfaces.js";
 
 import { createTracingContext } from "./tracingContext.js";
-import { state } from "./instrumenter-state.js";
+import { state } from "./state.js";
 
 export function createDefaultTracingSpan(): TracingSpan {
   return {
