+++ sdk/core/core-tracing/src/state-cjs.cts
