+++ sdk/core/core-tracing/src/state.ts
@@ -4,11 +4,11 @@
 import { Instrumenter } from "./interfaces.js";
 // @ts-expect-error The recommended approach to sharing module state between ESM and CJS.
 // See https://github.com/isaacs/tshy/blob/main/README.md#module-local-state for additional information.
-import cjsState from "../commonjs/instrumenter-state.js";
+import cjsState from "../commonjs/state.js";
 
 /**
  * Defines the shared state between CJS and ESM by re-exporting the CJS state.
  */
 export const state = {
-  instrumenterImplementation: cjsState.instrumenterImplementation as Instrumenter | undefined,
+  instrumenterImplementation: cjsState.state.instrumenterImplementation as Instrumenter | undefined,
 };
